pyproject.toml
[tool.poetry]
name = "myplaylist"
version = "0.1.0"
description = ""
authors = ["RustamSultanov <rustamelutingol@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.6"
django = "=2.0"
django-audiofield = "^0.10.0"
psycopg2 = "^2.7"
django-debug-toolbar = "^1.11"
django-environ = "^0.4.5"
psycopg2-binary = "^2.7"
Pillow = "^5.4"
gunicorn = "^19.9"
django-phonenumber-field = "^2.3"
django-tz-detect = "^0.2.10"
phonenumbers = "^8.10"
django-registration = "^3.0"

[tool.poetry.dev-dependencies]
pytest = "^3.0"

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
